Split out LXDContainerDirectories
Split out LXDContianerDirectories and LXDContainerUtils
