+++ nova_lxd/nova/virt/lxd/container_config.py
@@ -33,6 +33,7 @@ import six
 from nova_lxd.nova.virt.lxd import container_client
 from nova_lxd.nova.virt.lxd import container_image
 from nova_lxd.nova.virt.lxd import container_utils
+from nova_lxd.nova.virt.lxd import utils as container_dir
 from nova_lxd.nova.virt.lxd import vif
 
 _ = i18n._
@@ -47,7 +48,7 @@ LOG = logging.getLogger(__name__)
 class LXDContainerConfig(object):
 
     def __init__(self):
-        self.container_dir = container_utils.LXDContainerDirectories()
+        self.container_dir = container_dir.LXDContainerDirectories()
         self.container_client = container_client.LXDContainerClient()
         self.container_image = container_image.LXDContainerImage()
         self.container_utils = container_utils.LXDContainerUtils()
