+++ nova_lxd/nova/virt/lxd/container_utils.py
@@ -17,7 +17,6 @@
 from nova import exception
 from nova import i18n
 from nova import utils
-import os
 
 import container_client
 from oslo_concurrency import processutils
@@ -266,77 +265,3 @@ class LXDContainerUtils(object):
         except processutils.ProcessExecutionError as e:
             err = six.text_type(e)
         return err
-
-
-class LXDContainerDirectories(object):
-
-    def __init__(self):
-        self.base_dir = os.path.join(CONF.instances_path,
-                                     CONF.image_cache_subdirectory_name)
-
-    def get_base_dir(self):
-        return self.base_dir
-
-    def get_instance_dir(self, instance):
-        return os.path.join(CONF.instances_path,
-                            instance)
-
-    def get_container_rootfs_image(self, image_meta):
-        return os.path.join(self.base_dir,
-                            '%s-rootfs.tar.gz' % image_meta.get('id'))
-
-    def get_container_manifest_image(self, image_meta):
-        return os.path.join(self.base_dir,
-                            '%s-manifest.tar' % image_meta.get('id'))
-
-    def get_container_metadata(self, image_meta):
-        return os.path.join(self.base_dir,
-                            '%s-lxd.tar.xz' % image_meta.get('id'))
-
-    def get_container_rootfsImg(self, image_meta):
-        return os.path.join(self.base_dir,
-                            '%s-root.tar.gz' % image_meta.get('id'))
-
-    def get_container_configdrive(self, instance):
-        return os.path.join(CONF.instances_path,
-                            instance,
-                            'config-drive')
-
-    def get_console_path(self, instance):
-        return os.path.join(CONF.lxd.root_dir,
-                            'containers',
-                            instance,
-                            'console.log')
-
-    def get_container_dir(self, instance):
-        return os.path.join(CONF.lxd.root_dir,
-                            'containers')
-
-    def get_container_rootfs(self, instance):
-        return os.path.join(CONF.lxd.root_dir,
-                            'containers',
-                            instance,
-                            'rootfs')
-
-    def get_container_rescue(self, instance):
-        if self.is_lvm(instance):
-            return os.path.join(CONF.lxd.root_dir,
-                                'containers',
-                                instance)
-        else:
-            return os.path.join(CONF.lxd.root_dir,
-                                'containers',
-                                instance,
-                                'rootfs')
-
-    def get_container_lvm(self, instance):
-        return '%s/%s.lv' % (self.get_container_dir(instance),
-                             instance)
-
-    def is_lvm(self, instance):
-        try:
-            if os.path.exists(os.readlink(
-                self.get_container_lvm(instance))):
-                return True
-        except Exception:
-            return False
