+++ nova_lxd/tests/test_container_config.py
@@ -20,13 +20,13 @@ from nova import exception
 from nova import test
 
 from nova_lxd.nova.virt.lxd import container_config
-from nova_lxd.nova.virt.lxd import container_utils
+from nova_lxd.nova.virt.lxd import utils as container_dir
 from nova_lxd.tests import stubs
 
 
 @ddt.ddt
 @mock.patch.object(container_config, 'CONF', stubs.MockConf())
-@mock.patch.object(container_utils, 'CONF', stubs.MockConf())
+@mock.patch.object(container_dir, 'CONF', stubs.MockConf())
 class LXDTestContainerConfig(test.NoDBTestCase):
 
     def setUp(self):
